one runner for main/renderer processes

+++ packages/rpc/src/RPCProcess.js
@@ -56,7 +56,7 @@ export default class RPCProcess<Methods> {
     throw new Error('not implemented');
   }
 
-  async start() {
+  async start(): Promise<void> {
     this._ipc.config.id = this.serverID;
     this._ipc.config.retry = 1500;
     this._ipc.config.silent = true;
@@ -82,7 +82,7 @@ export default class RPCProcess<Methods> {
 
   stop() {
     this.server && this.server.stop();
-    if (this._subprocess) {
+    if (this._subprocess && this.isAlive) {
       try {
         process.kill(-this._subprocess.pid, 'SIGKILL');
         // eslint-disable-next-line no-empty
@@ -135,8 +135,6 @@ export default class RPCProcess<Methods> {
     }
   }
 }
-
-const RPC_PROC = new Map();
 
 const getBabelNodeBin = () =>
   path.resolve(__dirname, '../../../node_modules/.bin/babel-node');
